init-com

+++ tests/test_dynamics.py
@@ -62,7 +62,7 @@ def test_clip(
             step_action = [env.action_space.sample() for _ in range(batch_size)]
             step_action = torch.tensor(step_action).unsqueeze(1).float()
             step_action = step_action.repeat(1, num_ensemble, 1)
-            step_obs, reward, done = network.step(step_obs, step_action)
+            step_obs, reward, done = network.step(step_obs.float(), step_action)
 
             assert (reward_min <= reward).all() and (
                 reward <= reward_max
